+++ src/nectarchain/makers/component/__init__.py
@@ -1,14 +1,14 @@
-from .chargesComponent import ChargesComponent
+from .charges_component import ChargesComponent
 from .core import ArrayDataComponent, NectarCAMComponent, get_valid_component
-from .FlatFieldSPEComponent import (
+from .flatfield_spe_component import (
     FlatFieldCombinedSPEStdNectarCAMComponent,
     FlatFieldSingleHHVSPENectarCAMComponent,
     FlatFieldSingleHHVSPEStdNectarCAMComponent,
     FlatFieldSingleNominalSPENectarCAMComponent,
     FlatFieldSingleNominalSPEStdNectarCAMComponent,
 )
-from .gainComponent import GainNectarCAMComponent
-from .PedestalComponent import PedestalEstimationComponent
+from .gain_component import GainNectarCAMComponent
+from .pedestal_component import PedestalEstimationComponent
 from .photostatistic_algorithm import PhotoStatisticAlgorithm
 from .photostatistic_component import PhotoStatisticNectarCAMComponent
 from .spe import SPECombinedalgorithm, SPEHHValgorithm, SPEHHVStdalgorithm
